requirements.txt
bottlenose==1.1.2
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
pkg-resources==0.0.0
psycopg2==2.7.3
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.1
six==1.10.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
